+++ examples/00-quickstart/wgpu_basic.rs
@@ -311,6 +311,7 @@ impl AppWindow {
                 depth_stencil_attachment: None,
                 timestamp_writes: None,
                 occlusion_query_set: None,
+                multiview_mask: None,
             });
 
             // Call new_frame before rendering