+++ examples/01-renderers/sdl3_wgpu.rs
@@ -243,6 +243,7 @@ fn main() -> Result<(), Box<dyn Error>> {
                 depth_stencil_attachment: None,
                 timestamp_writes: None,
                 occlusion_query_set: None,
+                multiview_mask: None,
             });
 
             renderer.new_frame()?;