+++ examples/02-docking/multi_viewport_wgpu.rs
@@ -239,6 +239,7 @@ impl AppWindow {
                 depth_stencil_attachment: None,
                 timestamp_writes: None,
                 occlusion_query_set: None,
+                multiview_mask: None,
             });
             drop(rpass);
 
@@ -311,6 +312,7 @@ impl AppWindow {
                 depth_stencil_attachment: None,
                 timestamp_writes: None,
                 occlusion_query_set: None,
+                multiview_mask: None,
             });
 
             self.renderer.new_frame()?;