+++ examples/04-integration/ime_debug.rs
@@ -296,6 +296,7 @@ impl AppWindow {
                 depth_stencil_attachment: None,
                 timestamp_writes: None,
                 occlusion_query_set: None,
+                multiview_mask: None,
             });
 
             self.imgui.renderer.new_frame()?;