+++ examples/imguizmo_quat_basic.rs
@@ -853,6 +853,7 @@ impl AppWindow {
                 depth_stencil_attachment: None,
                 timestamp_writes: None,
                 occlusion_query_set: None,
+                multiview_mask: None,
             });
 
             self.imgui.renderer.new_frame()?;
@@ -1276,6 +1277,7 @@ impl AppWindow {
                 }),
                 timestamp_writes: None,
                 occlusion_query_set: None,
+                multiview_mask: None,
             });
             rpass.set_pipeline(&self.scene.pipeline);
             rpass.set_vertex_buffer(0, self.scene.vbuf.slice(..));