+++ examples/imnodes_basic.rs
@@ -1045,6 +1045,7 @@ impl AppWindow {
                 depth_stencil_attachment: None,
                 timestamp_writes: None,
                 occlusion_query_set: None,
+                multiview_mask: None,
             });
 
             let draw_data = self.imgui.context.render();