+++ extensions/dear-implot-sys/src/bindings_pregenerated.rs
@@ -9827,6 +9827,16 @@ unsafe extern "C" {
     );
 }
 unsafe extern "C" {
+    pub fn ImPlot_Annotation_Str0(
+        x: f64,
+        y: f64,
+        col: ImVec4_c,
+        pix_offset: ImVec2_c,
+        clamp: bool,
+        fmt: *const ::std::os::raw::c_char,
+    );
+}
+unsafe extern "C" {
     pub fn ImPlot_AnnotationV(
         x: f64,
         y: f64,
@@ -9844,6 +9854,9 @@ unsafe extern "C" {
     pub fn ImPlot_TagX_Str(x: f64, col: ImVec4_c, fmt: *const ::std::os::raw::c_char, ...);
 }
 unsafe extern "C" {
+    pub fn ImPlot_TagX_Str0(x: f64, col: ImVec4_c, fmt: *const ::std::os::raw::c_char);
+}
+unsafe extern "C" {
     pub fn ImPlot_TagXV(x: f64, col: ImVec4_c, fmt: *const ::std::os::raw::c_char, args: va_list);
 }
 unsafe extern "C" {
@@ -9851,6 +9864,9 @@ unsafe extern "C" {
 }
 unsafe extern "C" {
     pub fn ImPlot_TagY_Str(y: f64, col: ImVec4_c, fmt: *const ::std::os::raw::c_char, ...);
+}
+unsafe extern "C" {
+    pub fn ImPlot_TagY_Str0(y: f64, col: ImVec4_c, fmt: *const ::std::os::raw::c_char);
 }
 unsafe extern "C" {
     pub fn ImPlot_TagYV(y: f64, col: ImVec4_c, fmt: *const ::std::os::raw::c_char, args: va_list);